Turn XHR into a module
Rename XMLHttpRequestProgressEvent and expose it under
window.ProgressEvent, as this is the right name of the Web API (1).
Also move it to DOMEvent.js, as it is not specific to XHR.

Remove _HttpRequest type, as it is now private to the module and we
don't want to test and expose internals.

[1] https://developer.mozilla.org/en-US/docs/Web/API/ProgressEvent/ProgressEvent

Change-Id: I75df3bcafc3f596abe540cfa5b53a064fcfaa162