Add basic integration tests, revise main.js
Since all tests have now been transformed to unit tests, the tabris
module that is created during the build is not tested anymore.

Add a `verify` step to the build that runs after building the tabris
module and verifies that it is correctly assembled and exports the
expected API.

We can later extend this test to also include tests for the native
bridge.

Add some missing exports to window and tabris objects in main.js.
Reorder statements for clarity.

Change-Id: Ia39a42ae7af881c140b02d2633f49f7e48fb5291