Turn UI into a module
Leave tabris._UI intact, as it is used by several widget Jasmine tests.
This private property should be removed from tabris when it is not used
anymore.

Change "instance with a page ShowPreviousPage event closes page" test to
"instance with a page fails when closing the single page". It didn't
fail with Jasmine, although it should have failed, because we close the
last page on the UI.

Change-Id: Ib99fff39aa2651b5083e7ae68a7469684948e6f5