Rename NativeBridgeSpy to ClientStub
The tabris object holds the fields `_client` and `_nativeBridge`.
The `_client` object is provided by the native client and it's wrapped
by the native bridge.

What we use to call "NativeBridgeSpy" does not simulate the native
bridge, but the `_client` object. Moreover, it's a stub rather than
a spy as it does not wrap an existing object but acts as a stands in.

Rename "NativeBridgeSpy" to "ClientStub" and the instance variables
from "nativeBridge" to "client".

Change-Id: Iad13914d5e1cfdd5021d1206359423f32073cb24