Implement NavigationView
NavigationView serves as the new container for instances of Page.
It has a read-only "stack" property which is an object representing the
current stack of pages. It can be manipulated using an array-like API.

This implementation does not yet consider the case that a page is
disposed while on the stack. There is also no way yet to access pages on
the stack except the first and the last, or use NavigationView with
PageSelector.

Adjust one snippet to showcase the usage of the new API.

Documentation will be updated separately once the API can be considered
final.

Change-Id: Ibbc2ede68985ab8271afd6dccd1370dd65cda12a