Add documentation for NativeObject
Instead of documenting inclusion of "Events API" and "Properties API",
let widgets inherit from NativeObject that exposes these APIs.

This is easier to understand and matches the actual implementation.

Since WidgetCollection does not inherit from NativeObject but
implements a couple of events and property related methods itself,
document these methods in the context of WidgetCollection.

In the descriptions, omit the statement "supports chaining" as
this is obvious by the return type "this".

Change-Id: I76b5445b8e51ba413f5d840a3ce56efe3a458337