Use ContentView as container in most snippets
It is not necessary to create a page container anymore in Tabris.js 2.x.
Instead the statically available ContentView can be used.

Snippets which rely on pages and the toolbar have not be modified, i.e.
action, page and ui snippets. Those will be adjusted later as API is
published.

Change-Id: Iecb8b7ef63106ef0ef9e2ba449f7ddebec41cfdf