Migrate remaining jasmine tests to mocca
Adjust Module.js and bootstrap.js to work with the global object
exposed as `global` instead of `window` for compatibility with Node.js.
We should adjust the clients to do so as well.

The boot modules creates global objects tabris and document.
Run the boot tests in a separate task to avoid effects on the
tabris test suite.

Remove dependency to jasmine.

Change-Id: I7ab8f9e65c0a39bd47fc7426dd5dea2ffbc26b00