Revise dispose checks in NativeObject
Consistently throw an Error when calling any of the _nativeXxx()
methods on a disposed object. After an object has been destroyed,
any subsequent operations would result in client errors.

When calling public methods get() and set() on a disposed object,
print a warning to indicate the problem. Application code should
not call these methods after dispose, however, it's not critical
and throwing would be too drastic.

On a disposed object, let get() return undefined and set() be a
no-op. Delete the property store on dispose to free memory and to
ensure that also cached properties would not be returned anymore.

Complete and cleanup tests, suppress console output.

Change-Id: I452d0a4c5b150a05707bb95b3d140825c05d9078