[doc] Include basic snippets in API and widgets
Include the corresponding snippet in each widget and API documentation
article when a snippet is available. The snippet will be shown before
the "See also" section. Other snippets, e.g. property examples, may be
included at a later stage.

Restore snippet descriptions for snippets to be shown as comments
inlined on top of each snippet. They were removed in @5cb59e0.

Remove snippet and example source location configuration from Gruntfile
and hard code it, as we don't need to configure it and copy-examples and
generate-doc modules are not generic.

Change-Id: I2397e58799d7566343c54b9aef287f28028a87a4