Invert delegation of setters and getters
When generating properties, instead of letting the properties delegate
to the generic get and set methods, let the generic methods delegate
to the generated properties.

This allows to override specific properties in subclasses without
having to override get or set.

Log a warning when trying to set an unknown property. The getter
simply returns undefined for missing properties.

In the Picker widget, the order of properties matters. Instead of
overwriting _setProperties method, introduce a method
_reorderProperties that can be overwritten by clients.

Change-Id: I625f40d14193ee2c9135daf648c9cc3e2012d4a0