Update eslint, adjust eslint config to ES6
Since the sources are now transpiled, change the environment to "es6"
in the eslint config. This includes support for Promise, ArrayBuffer,
etc., so these globals don't have to be declared anymore.

Remove other global variables that should not be used in the source
code, move them to snippets and examples, where they are allowed.
This helps us keep the source code clean.

Change-Id: I0fbf0363e4ff1ad43cc40a1fbc5fdca40e7fbc99