[ES6] Use let instead of var
After adding a transpile step, we can now use ES6 syntax.

Add the rule "no-var" to ESLint config. Disable the rule for snippets,
examples, and also for src/boot/ as this code is not yet transpiled.

Run eslint the --fix option to adjust the code.
As a side-effect, this also adjusted the Gruntfile and a grunt task.

Change-Id: Icf90d91fd17f51eca5cce11b3e0bcade3e703e8f