[ES6] Use arrow functions where appropriate
Add the "prefer-arrow-callbacks" ESLint rule [1] that suggests the use
of arrow functions as callbacks where possible (i.e. `this` is not used
in the function body).

Disable this rule in test/, as Mocca discourages the use of arrow
functions for `describe` and `it`. Even if we disagree with this
recommendation, the changes would be very broad and should go into
a separate change.

Run eslint with the --fix option to adjust the code.

Also add the rule "arrow-body-style" to "as-needed" to disallow

    () => { return expr; }

in favor of

    () => expr

This change required some manual adjustment.

[1] http://eslint.org/docs/rules/prefer-arrow-callback

Change-Id: I4f514f7390f8907257d11d60847611e58ab8aca1