Simplify copy-examples build script
Use syncExec to run `npm install <module>` instead of the hardly
documented npm API. It's simpler and results in shorter and clearer
code without passing callbacks. There is no point in asynchronous
operations in a build script anyway.

The syncExec method throws if the command does not terminate
successful, which would cause the grunt task to fail.

Remove dependencies npm and async.

Change-Id: I8f9e08f314a1ea2a4be8e81fedfc126b72ccac09