[doc] Use "extends" instead of "include"
With the inclusion of the Properties implementation in NativeObject,
Tabris.js now has a clear type hierarchy without multiple inheritance.

Reflect this structure in the type declarations and in the generated
documentation by replacing the term "inherit" with "extends".

In the doc generator, merge the functions resolveIncludes and
renderIncludes. The "resolve" step has become obsolete.

Change-Id: I3ed03927d5ced607001b2b4a22aa3e9cb84edd74