Fix error with undefined tabris.load function
A version check in Widget.js accesses the tabris.load function at
module loading time, before the global tabris object is published.

Move this check into a function that is executed at runtime, without
using tabris.load() which we want to get rid of in the long run.
Cache the result to avoid unnecessary context switches.

Change-Id: I48dc26b9aa95a8e948308a1b5cbc035d06ded8be