Fix items property in CollectionView
CollectionView overrides the `set()` function to trigger a deferred
reload function. This has no effect when the properties are set
directly.

Execute this function on flush instead of overriding set().

Adjust the snippet collectionview-refreshenabled to verify. This is
the only snippet that sets the items property outside of create.

Fix #1028

Change-Id: Iac32e95c624136e9f4b84710137ed7999774c86b