Add property "tabMode" to TabFolder
The property can have the values "scrollable" and "fixed" (default).
The mode "fixed" reassembles the previous behavior.

The "tabMode" controls how the tabs make use of the available
horizontal space. Setting the `tabMode` to "fixed" makes the
tabs span the entire available space. In case of a very wide
`TabFolder` the "fixed" mode centers the tabs. The mode
"scrollable" left aligns the tabs and allows to scroll the tabs
if there are more tabs than would fit in the available space.

Available on Android only.