Support native pbkdf2 function
To make password hashing secure, a hash function must be applied
multiple times. With common iteration counts of >> 1000, this is
too slow an operation to be performed in plain JS.

Introduce a (currently undocumented) object pkcs5 on tabris that
provides a pdbdf2 function with a fast native implementation.

Since even a native implementation can take a perceivable amount
of time, make this function asynchronous, i.e. return a Promise.

Transfer the salt as a plain array of numbers (0..255) to the client
as iOS can't currently handle typed arrays without pre-processing.

Change-Id: Ib21c051937abe84da04486c3e0f27ee32327efd7