Support arraybuffer responseType in XHR
Revise the implementation of the properties responseType, responseText,
and response according to the latest version of the spec [1].

Rename the scope property responseText to responseData, it can now also
contain other types than strings.
Trust the client to provide the correct data, omit type checks in
response and responseText getters.

[1] https://xhr.spec.whatwg.org

Change-Id: I1d11180d07a3f85dee3376bb2331e819c0c14238