Support Array includes function
The use of `indexOf()` to check for array containment is often
confusing, especially as it involves comparing with -1 (`!== -1` vs.
`=== -1`). Using the new `includes()` function is more expressive.

Unfortunately, this function is not yet supported on iOS 8 (on iOS 10
it is).

Use the third-party babel plugin "array-includes" that rewrites
`array.includes(x)` to `array.indexOf(x) !== -1` to allow us to use
array.includes() in our code.

Adjust util.js for a start.

Change-Id: I0e980fd8d58fcdd6553736391d3464bbf025c2ec