Let the parent function return null by default
When a widget has no parent, the parent() function currently
returns `undefined`.

I think we should generally avoid `undefined` in favor of `null` in
the API. We shouldn't end up with a mix of both, and `null` is more
common to represent "no value".

Change the return value to `null` when there is no parent.

Change-Id: I93c2c0e460b35cf1c2cf497e1949e3788b505474