Enable widget append() / dispose() API for pages
Allow adding pages to and removing pages from a NavigationView using
standard widget API:

- Use `page.appendTo(navigationView)` or `navigationView.append(page)`
  to add a Page to a NavigationView.
- Use `page.detach()` to remove a page from a NavigationView.
- Use `page.dispose()` to remove a page from a NavigationView and
  dispose it.
- When a page is removed from a NavigationView's page stack, any pages
  on top of it are removed as well.

Provide a pages() function on NavigationView that returns a
WidgetCollection of all pages. This function is similar to `children()`
but it filters out widgets of other types such as `Action`.

Provide a temporary stack property that delegates to the new API to
allow for a smooth migration of already existing apps.

Change-Id: I69cfb38a959567a41301bc14b1d15ccda2859b87