Add a method to remove a widget from its parent
Widgets in tabris.js are already reparentable, but they can only be
removed from their parent by appending them to another (limbo) parent.

Detaching a widget can be useful, e.g. to rescue it on dispose and to
re-use it later.

Add a `detach()` method to remove a widget from its parent.

A `detach` method also exists in jQuery. The term "detach" seems more
descriptive than "remove" as it makes it clear that the widget is
transferred into a "detached" state.

Change-Id: I74ba505dd6ca0f1c8e2036ec24349a5bf26484f5