Implement NativeObject and Widget as ES6 classes
ES6 classes make it easier to handle inheritance by providing a clear
prototype chain, super calls, etc.

Use ES6 classes for the base classes NativeObject and Widget as a
preparation for switching to classes entirely.

Consider NativeObject abstract, disallow creating instances directly.

Move 'tabris.js not started' check to the NativeObject constructor.

Wrap Events in a class to allow extending it. It can't be added to the
prototype, otherwise the `_listen()` method in NativeObject would be
overridden by the one in Events.

Change-Id: Id7f07b87cc1c80b65ee7e6fcd73a794b37316f99