Use ES6 classes for widgets
As an incremental step towards using ES classes, define all widgets
as classes that extend an intermediate class created by Widget.extend
from a config object.

Move all methods that are not part of the configuration to the class
body. Use super calls instead of this._super. Call super with proper
parameters instead of applying arguments in order to maintain the
function signature.

Another advantage of this approach is that widgets constructors have
the proper widget name, i.e. `widget.constructor.name` returns the
class name if not mangled by a minifier.

Change-Id: I52942cde9aaa9a2fe29686eb4ffaa6fa7bb7e9ed