Replace extend and clone with Object.assign
Use the ES6 function `Object.assign()` instead of the util function
`extend()` - `extend` is shorter, but `Object.assign` is standard.
The functions are equivalent except for the inclusion of non-enumerable
properties - which doesn't make a difference for our use cases.

Also replace the util function `clone` with `Object.assign({}, ...)`.

Include a polyfill because Object.assign is not yet supported on iOS 8.
Use the npm module "es6-object-assign" which is minimal and follows the
polyfill outlined on MDN [1].

[1] https://www.npmjs.com/package/es6-object-assign
[2] https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/assign

Change-Id: I65b4da0c183ea79137d11c870c2c99affcf5ca8c