Use ES6 classes for native objects
Define all subtypes of NativeObjects as classes that extend an
intermediate class created using extend from a config object.

Move all methods that are not part of the configuration to the class
body. Use super calls instead of this._super.

Remove support for _super from NativeObject.

Do not copy arbitrary members from the config object to the prototype
anymore, instead throw an error for unsupported config objects. This
will help to detect outdated third-party widgets.

Change-Id: Id62756064b153463e1aa9b2b84a4a17d75458d78