Skip uglify step for tabris.js
Uglify.js does not handle ES6 input. This required transpiling for-of
statements even though they are supported by the client platforms
(commit 90b94160). Unfortunately, the transpiled for-of code contains
Symbol.iterator, which is also not supported on iOS <10.

While we move on to newer platform versions that require less
transpiling, this will become even more of an issue.

Moreover, Uglify mangles constructor names, which prevents us from
using them in the selector API.

Leaving out the minification does not make the loading significantly
slower, so skip it for now to resolve these problems.
Keep it for boot.js and polyfill for now.

We should consider using an ES6-aware compressor instead, such as the
one developed by the Babel project [1].

Add checks for constructor names to the verification of the build
result.

https://babeljs.io/blog/2016/08/30/babili

Change-Id: I7a099709f55c1a57803e8383ffce69e51bc47085