Improve drawer-pages snippet
Use 'icon' page field instead of 'image', because as of Tabris.js 2.x,
Page.image will be shown in the navigation view title area and in the
snippet this is not expected.

Rename button 'Create another page' -> 'Create page in drawer' to make
it more clear that the active page won't change.

Set navigation view property 'animated' to false. This will solve the
unexpected animation on iOS when the stack is cleared and a new page
gets opened (property supported on Android, iOS implementation TBD).

Change-Id: I7bb0b50c383f57d68885d3857a5e0cedd098581e