Add popTo call for NavigationView protocol
In order to play the appropriate animation on page transitions, the
iOS client needs a clue how many pages will be popped off the stack.

Call a method `popTo` with a parameter that contains the page to
reveal, or null if the entire page stack is cleared. This replaces
the stack_clear method.

This call will be sent to the client before one or more pages are
removed from the stack.

Change-Id: I35fc0125a615f8d2d6ac1029119d3219c2eb4b0e