Remove support for _initProperties
Instead of passing an `_initProperties` object to `extend()`,
widgets can override the internal `_create` method.

Replace _initProperties in Picker, which is the only widget that
used this feature. In this special case, the property selection
conflicts with the property selectionIndex, so set the initial
property selectionIndex only when there is no selection property
defined.

Remove support for _initProperties.

Change-Id: I66361149d9dfd25a89ec10dba3208fad6ab03b35