Remove _cid, _name, and _type from constructor
The members `_cid` and `_name` are only accessed within the extend()
function. They don't have to be kept statically. Remove them from the
constructor.

The member `_type` is used in the `_create()` method. Pass it to this
method directly. Changing the signature of _create() will allow custom
widgets to specify the type in the constructor, without using extend(),
e.g.:

    constructor(properties) {
      super();
      this._create('my.type', properties);
    }

Cleanup NativeObject: replace generic staticMembers map with specific
values.

In create(), do not fall back to the widget name (this.type) for the
type passed to _nativeCreate() anymore. The native type should not be
a matter of guessing.

Change-Id: Ib182522a4bae18934df8311364d9a086c35f7b00