Include babel minifier
Include the minifier preset provided by the babel project [1].
This setup reduces the size of the tabris.js by roughly one third
(179k -> 114k).

Disable "unsafe" modules as well as the simplify module which leads
to an "o is not defined" error when loading the tabris module.

Set the option "keepFnName" for the mangle module to preserve
constructor names.

Format JSON.

[1] https://github.com/babel/babili/tree/master/packages/babel-preset-babili

Change-Id: I13a47741892aaa8615fa8bd92c2decca217d99b1