Modify event listener API for Widget events
In an effort to change the signature of event listeners to accept
a single event object instead of separate parameters, adjust events
on NativeObject and Widget to this pattern.

This includes gesture and touch events, animation events, resize and
dispose.

Introduce generic interfaces for type definitions, add typing tests
in Widget.test.js. Also introduce a basic event interface to inherit
from. To distinguish it from DOM events, call it `EventObject` for now.

Adjust documentation.

Change-Id: Iaf13fb1cb42c738caea2ee32da80bb1dcfee6dcd