Update documentation of event-related methods
Adjust the documentation of `on`, `once`, and `off` to the new API,
i.e. only a single event object is passed to listeners. Rename the
first parameter `event` to `type` to avoid confusion with the event
object. `type` is the term used in the DOM events documentation [1].

Also change the documentation for `trigger` and declare only a single
parameter. Although the trigger method still passes all parameters to
listeners, we should not advertise this anymore.

The listen function should be adjusted in a separate change.

[1] https://developer.mozilla.org/en-US/docs/Web/API/EventTarget/addEventListener

Change-Id: Ibe20a78a2d425e12b6f4c746b28523e14bbe344b