Skip additional parameters in trigger method
With the new events API, listeners should always receive exactly one
parameter, the event object.

However, the trigger() function passes all parameters after the first
one to the listeners. This does not fit with definition of the on(),
off(), and once() methods.

Let the trigger function pass only one parameter to the listeners.
If called with only the type, pass an empty object to the listeners.

Adjust and cleanup tests:
- rename "callback" to "listener"
- omit "should" in test names for brevity
- follow the build-operate-check pattern to improve readability

Change-Id: I7da95d184878ff2646261c50c2f6fe4f7859aba3