Remove replicated event methods in type definition
Inherit generic event methods on(), off(), once(), and trigger() from
NativeObject instead of replicating these definitions for every type.
This reduces the size of the definition file by 2/3.

This change temporarily removes support for auto-suggestion of event
names. Definitions of specific event names should be added again, but
with a different approach that does not involve this redundancy.

Change the type of the event object to `any`, as `Object` does not
allow access to arbitrary properties.

Change-Id: I8b71c1af0a98690852d9994b68cfa1b6d71156db