Provide static method to define properties
In an effort to replace `NativeObject.extend()`, provide a static
method `defineProperties()` to define properties on a given prototype.

Keep the logic of property definitions, normalization, etc. intact.
Instead of a `$properties` map, store a separate `$prop_name` member
on the prototype to allow for inheritance.

Implicitly test the method by using it instead of Widget.extend for
property-related tests in NativeObject.test.js.

Adjust Animation.js as an example for a type that does not use
NativeObject.extend().

Change-Id: I22923f038c890b66d8203423a81c544d71bdbd11