Rename button property "selection" to "checked"
On stateful button widgets (CheckBox, Switch, etc.), the property
name `selection` is rather unspecific and ambiguous. Rename it to
`checked` which is more obvious and also matches with the Android
API.

In the `select` event, rename the parameter `selection` to `checked`.
The event itself should be renamed in a separate change.

Also adjust snippets and examples.

Change-Id: I1b168fac1540085ad1b7109d8cc606279abed419