Replace NativeObject.extend() with reduced version
This method is not used by tabris-js anymore, but it is still useful
for creating custom objects in ES5, e.g. in cordova plugins that
don't include a transpiler.

Since NativeObject is implemented as a class, prototype inheritance
is very complicated. The NativeObject constructor can now neither be
called directly, nor can it be called as a function using call().

Replace NativeObject.extend() with a reduced version that only
accepts a native type and optionally a super type.

Change-Id: I7bad4f0026f394f8db1a4f335e88614265dea84b