Include missing type definitions for globals
Remove the libraries "dom" and "scripthost" from the tsconfig file.
Tabris.js only provides a subset of these libraries.

Include a d.ts file for the global timer functions provided by
Tabris.js.

Adjust included type definition of fetch to Tabris.js, which does
not provide the types Blob and FormData.

Include the type ImageData in tabris instead of expecting it in the
global scope.

Change-Id: Ia6ec0f69b0779c53129483c5ec7e5d8da840c56a