Remove animation events
The events `animationstart` and `animationend` are problematic,
because listeners must be added before calling `animate()`.
Moreover, listeners can not be added for a specific animation.
If there are multiple animations, they need to be marked and then
filtered in the listener code.

Since the `animate()` function now returns a promise, these events
are not needed anymore.

Remove support for both events.
Ignore the native `Start` event which is not needed anymore.

Change-Id: Id6373eebadd2371dd31a3e99dc11077adeaf2903