Simplify property definition syntax
Remove the surrounding `access` object for `get` and `set` functions.
This object had once been introduced to allow for preconfigured modes
such as `readonly`, however, this has never happened. Since property
definitions are internal API, we should keep it simple.

Keep support for shorthand definitions that only contain a type, but
only if the shorthand is a string, e.g:

    foo: 'number'

Remove the obsolete parameter `options` from property setters. This
parameter is not supported anymore.

Change-Id: Ibabfcb68efd6d8f05b482034780a59a29702967f