Extract method for mocking tabris in tests
There's a common code snippet to mock tabris in most tests.
This code is copied from test to test and it clutters up the
test setup with internals.

Extract this mocking code to a module of its own.

Provide a method mockTabris() in the test facade and use it in tests.

Remove obsolete unused file `tabris-init.js`.

Change-Id: I1e565f2efdea9aca25f789e53e8d04831f6536f9