Remove cross-env module
The call to cross-env is broken with the most recent version of this
module because the binary has been moved to a different location.

This module is used only to set the environment variable BABEL_ENV
which is not supported inline by some Windows shell.

Remove cross-env and set the environment variable directly in the exec
task configuration.

Change-Id: I272f86b19ddf05fb6c02d76f3c1661ae67a7006b