Remove WebSocket example from app examples tab
The WebSocket example is not self-contained. It requires a server
running and a modification of the source code, so it cannot be used
from the developer app on its own.

Having a 'title' in the package.json is the condition for an example
to be shown on the examples tab of the developer app.

Change-Id: Ie7a8665132a12e6e5104cf251435be254bcfec43