Support removing hidden pages from NavigationView
The native clients support removing a page from a NavigationView
that is not the topmost page. This allows removing underlying pages,
e.g. when an action is completed, and so changing the target for
back navigation.

This change removes API exceptions for NavigationView and adjusts
its behavior to other widgets.

Adjust UI doc to the API change.

Fix #1124

Change-Id: I6aad81cdc18847df462234c7b4fcfbe75a6e0dfb