Return created CollectionView Cell to client
When the native client requests the creation of a new CollectionView
cell using `createitem`, return the cid of the created cell as result
of the _notify() call. This simplifies matters for the native clients
and makes the `additem` call obsolete.

The `additem` method is still called. It should be removed once all
clients have been adjusted.

Change-Id: Ica80da0dbf38fdf6c895406af75fa3ddc6d1c48d