Provide basic event properties
Equip all events fired by Tabris.js with a number of basic
properties that are compatible with DOM events.

Introduce base class EventObject that implements the read-only
properties `type`, `target`, and `timeStamp`. Also include a
`preventDefault` method and a corresponding `defaultPrevented`
property.

Do not expose the constructor yet, this class is only meant to
implement a consistent event interface.

Always use the internal method _trigger() to trigger events rather
than trigger() to ensure the event is wrapped in an EventObject.

In custom _trigger implementations, always return the result of the
super call to pass the event's defaultPrevented state as result to
the client.

Change-Id: Ifaaf272ccb24311f5653f7178b2a30a2916c1360