Introduce internal native text decode function
In order to fully support the fetch API, the response data must be
translated to a string, using the content type encoding. Since the
JS engines don't have built-in support for text encoding, and third-
party libraries would inflate the tabris module, a native method is
preferable.

Introduce a native object type `TextDecoder` with a single method
`decode()` that accepts an ArrayBuffer and an encoding. This method
is supposed to perform the encoding asynchronously and send an event
when ready.

Change-Id: If4f17703f90766dcfc0ff117bca94e723f7a5386