Introduce createItem property on CollectionView
Allow application code to create cells on demand by implementing
`createCell` instead of `initializeCell`. This also enables ES6
code to return subclasses of Cell.

Keep support for initializeItem, but mark it as deprecated. The
generator will be adjusted to render the deprecation warning in
a subsequent commit.

Implement both properties not to be mutually exclusive. If both
properties are provided, call initializeCell with the cell created
by createCell().

Adjust one snippet as an example. The remaining snippets and
examples should be adjusted in a separate commit.

Change-Id: I36aa63b9f54e0f766fbf8336705dccc246cfecd0