Implement fetch in Tabris.js, remove polyfill
The fetch API in Tabris.js is provided by including a polyfill [2].
This polyfill is based on XHR and other Web APIs such as Blob,
FormData, and FileReader. This implementation can't be used to
fetch binary data without some of these APIs being available.

Even if these types were implemented or faked, the polyfill would
always request the response data as Blob and use FileReader to parse
it, which appears too much overhead for Tabris.js.

Implement the fetch API [1] in ES 6, based on the polyfill code [2].
Instead of XHR, communicate directly with the native client using
the HttpRequest remote object. Use the asynchronous TextDecoder
provided by the native clients.

[1] https://fetch.spec.whatwg.org/
[2] https://github.com/github/fetch

Change-Id: Ie292cdac5791b8e93231e8f737ae584696d887d4