Separate x/y properties on pan event
Instead of a single property `translation`, provide separate
properties `translationX` and `translationY` to match with the
properties of the Widget's `transform` object. Since the event
is an object already, it shouldn't have unnecessary nested objects.

For consistency, also split `velocity` into `velocityX` and
`velocityY`.

Add the new properties without removing the old ones to remain
compatible with existing code and the existing clients.

Change-Id: Id027d9c2ce0ed68365fbbfa71d33e511ca622dcc