Prevent appending CollectionView cells
CollectionView cells are managed by the framework, they should not
be appended by application code.

Trying to append a cell to a CollectionView would set the native
parent property, which might confuse clients. It would also modify
the CollectionView's children list.

Prevent appending cells to a CollectionView and log a warning.

Change-Id: I902ca96fb0547df1ffebeccd2cfb7f7d819cb91a