Fix performance issue in GC
The GC renders operations in a different format for iOS.
This platform switch requests the platform from the client in every
single drawing operation, drastically slowing down Canvas operations.

Store the platform in the GC instance.

Another option was to store the platform statically, but this would
make those platform switches hard to test. Once the iOS client is
adjusted to the new format, these switches can be removed anyway.

Change-Id: Ia323de462e11f69340f85ba521fa6c06a760736f