Make common Event properties enumerable
Event properties are enumerable in the DOM, but they aren't in Tabris.
As an effect, they are not included in for..in loops and they are
omitted in console.log.

Make common event properties type, target, and timeStamp enumerable.
Explicitly exclude them from Object.assign.
Add a missing test for _triggerChangeEvent.

Change-Id: I80a2ee23ac710dcc61123a08bb15c8e1b62b5569