Add timeout option to fetch
There is a long discussion about a timeout option in the fetch spec [1].
Chances are that a property named `timout` will be included in the
config object.

Without this property, every reasonable app has to wrap the fetch
promise in a timeout wrapper. Tabris can provide a native timeout
implementation which is superior, as it actually aborts the running
request.

Proactively provide a this config property to make use of the native
timeout implementation.

[1] https://github.com/whatwg/fetch/issues/20

Change-Id: I88059e9ff306e4c15196ba97bd92cba58a04aa77