Fix layout on "read book" page in bookstore example
The text TextView was covering the entire screen leading to a layout
where the book text was vertically centered (due to the TextViews
behavior to center text) although there is not enough text to cover the
available space. With this change the TextView is now only top attached
leading to the expected behavior.