Create native instance of Device and App
In the native bridge communication, there is a concept of so-called
"service objects", i.e. native objects with a pre-defined cid that
are expected to exist on the client without having to be created.
For these objects, the cid refers to the type.

These service objects pose an exception from the rule that is not
really needed. This concept should be abandoned.

To move forward in this direction, create an instance of tabris.App
and tabris.Device, but keep the widget id the same for compatibility.

Change-Id: I997d3a19f16d3fa69bfd0f384781999ec1b77027