Create native instances of ClientStores
In the native bridge communication, there is a concept of so-called
"service objects", i.e. native objects with a pre-defined cid that
are expected to exist on the client without having to be created.
For these objects, the cid refers to the type.

These service objects pose an exception from the rule that is not
really needed. This concept should be abandoned.

To move forward in this direction, create native instances of
tabris.ClientStore and tabris.SecureStore, but keep the widget
ids the same as the type for compatibility.

Change-Id: I10077835aa1bbd5c0c690d11be463b1314b15054