New CollectionView API
Let the CollectionView API work on indexes instead of model items.
This gives the developer more control over binding arbitrary types
of collections to a CollectionView. It also allows for large virtual
collections that load their items on demand.

Allow for arbitrary widgets as CollectionView cells. Update cells
using a dedicated callback instead of the `change:item` event. This
makes the type Cell equivalent to Composite and this redundant.

- Replace the `items` property in CollectionView with `itemCount`.
- Adjust API to work with indexes instead of items.
- Allow `createCell` to return arbitrary types of widget.
- Introduce `updateCell` method to update a cell with a new item.
- Rename `itemHeight` to `cellHeight` for consistency.
- Remove type `Cell`.

Change-Id: I2a61fd0da23c8d28682c8d204443b6a958b4cc01