Add support for 'enterKeyType' on TextInput
The property 'enterKeyType' allows to specify the label or icon shown on
the confirmation key of the TextInput keyboard. The following values are
supported: 'default', 'done', 'next', 'send', 'search', 'go'. The key
press can be captured via the `accept` event.

Setting an `enterKeyType` other than `default` will change the key
behavior to not close the keyboard automatically. The developer is able
close the keyboard by removing the focus from the `TextInput`.

Change-Id: I39f9387b1d50678b50e1cf7c9aa0bcff3d05d456