Move version check into tabris module
The version check in the boot.js is executed after the tabris module
have been loaded. When the loading of the tabris module fails due to
incompatibilities in the protocol, the version check will be too late.

Move the version check into the tabris module and perform it right
after creating the App instance.

Skip the check when the clientVersion is undefined to prevent failure
during tests.

Change-Id: I78b4e655ae04ec11f1a444ec1ea989b12d9fe229