Rename change events to camelCase
In an effort to remove special characters from event names,
e.g. to use them in object keys or method names, replace the
colon in the names of change event with camel case.

Use the suffix "Changed" instead of the prefix "change" to avoid
having to change the case of the property name.

As an example, `change:text` becomes `textChanged`.

Use the past tense "Changed" instead of "Change" to clarify that the
event is triggered *after* the property change. We don't use past
tense in other event names yet, however, at some point, we'll have
to distinguish between before and after events. So a breaking change
should go into the right direction.

Change-Id: Ie5ae4a10d381f8cd2aee5c79baa348308af05d13