Remove hard-coded definitions from type definition generator
The parameter for the `set` method is a properties-interface that
depends on the class the method is called on. E.g. the set method on
Widget accepts a parameter of type `WidgetProperties`.
For this reason, the `set` method has to be re-defined for every class
that inherits it. This was previously achieved by adding the set method
explicitly to each class that is a descendant of NativeObject during
type definition generation.

This change removes these hard-coded method definitions for `set` and
instead introduces the `PropertiesObject` type. When generating the
methods for a class, the type definition
generator finds all methods that have a parameter of type
`PropertiesObject` in its parent classes and generates a new method
definition with the appropriate properties-interface type for the
parameter.

This makes it easier to add similar class dependent methods and
properties, like constructors, the `on` method and event-interfaces,
without increasing the complexity of the type definition generator.


Change-Id: I3ed8846f23cfe326ee01fb1ac742071b3bfbe1cf